reorganise files in pipeline folder
