+++ pipeline/models/prophet_model.py
@@ -9,7 +9,7 @@ from sklearn.metrics import (
     mean_absolute_percentage_error,
 )
 
-from pipeline.jobs.forecast_model import ForecastModel
+from pipeline.models.forecast_model import ForecastModel
 from utils import get_logger
